+++ admin/fuisAssegnatoTipoReadRecords.php
@@ -9,11 +9,15 @@
 
 require_once '../common/checkSession.php';
 
+$soloAttivi = $_GET["soloAttivi"];
+
 // Design initial table header
 $data = '<div class="table-wrapper"><table class="table table-bordered table-striped table-green">
             <tr>
-                <th>Nome</th>
-                <th>Modifica</th>
+                <th class="text-center col-md-8">Nome</th>
+                <th class="text-center col-md-2">Citrix</th>
+                <th class="text-center col-md-1">Attivo</th>
+                <th class="text-center col-md-1">Modifica</th>
             </tr>';
 
 $query = "	SELECT
@@ -21,13 +25,24 @@ $query = "	SELECT
 				fuis_assegnato_tipo.*
 			FROM fuis_assegnato_tipo
 			";
+if( $soloAttivi ) {
+	$query .= "WHERE fuis_assegnato_tipo.attivo is true ";
+}
 
-$query .= "order by nome";
+$query .= "ORDER BY attivo DESC, nome ASC";
 
 foreach(dbGetAll($query) as $row) {
+    $statoMarker = '';
+    if (! $row['attivo']) {
+		$statoMarker = '<span class="label label-danger">disattivato</span>';
+	}
+
     $data .= '
             <tr>
-                <td>'.$row['nome'].'</td>';
+                <td>'.$row['nome'].'</td>
+                <td>'.$row['codice_citrix'].'</td>
+                <td class="text-center">'.$statoMarker.'</td>'
+            ;
 	$data .='
                 <td>
                     <button onclick="fuisAssegnatoTipoGetDetails('.$row['local_id'].')" class="btn btn-warning btn-xs"><span class="glyphicon glyphicon-pencil"></button>
@@ -40,4 +55,3 @@ $data .= '
         </table></div>';
 echo $data;
 ?>
-