+++ admin/js/fuisAssegnatoTipo.js
@@ -5,8 +5,20 @@
  *  @license    GPL-3.0+ <https://www.gnu.org/licenses/gpl-3.0.html>
  */
 
+var soloAttivi=1;
+
+$('#soloAttiviCheckBox').change(function() {
+    // this si riferisce al checkbox
+    if (this.checked) {
+		soloAttivi = 1;
+    } else {
+		soloAttivi = 0;
+    }
+    fuisAssegnatoTipoReadRecords();
+});
+
 function fuisAssegnatoTipoReadRecords() {
-	$.get("fuisAssegnatoTipoReadRecords.php", {}, function (data, status) {
+	$.get("fuisAssegnatoTipoReadRecords.php?soloAttivi=" + soloAttivi, {}, function (data, status) {
 		$(".records_content").html(data);
 	});
 }
@@ -35,10 +47,14 @@ function fuisAssegnatoTipoGetDetails(id) {
 		},
 		function (data, status) {
 			var record = JSON.parse(data);
+            console.log(record);
 			$("#nome").val(record.nome);
+			$("#codice_citrix").val(record.codice_citrix);
+            $("#attivo").prop('checked', record.attivo != 0 && record.attivo != null);
 		});
     } else {
         $("#nome").val("");
+        $("#codice_citrix").val("");
     }
 	$("#update_modal").modal("show");
 }
@@ -46,7 +62,9 @@ function fuisAssegnatoTipoGetDetails(id) {
 function fuisAssegnatoTipoSave() {
     $.post("fuisAssegnatoTipoSave.php", {
         id: $("#hidden_record_id").val(),
-        nome: $("#nome").val()
+        nome: $("#nome").val(),
+        codice_citrix: $("#codice_citrix").val(),
+        attivo: $("#attivo").is(':checked')? 1: 0
     },
     function (data, status) {
         $("#update_modal").modal("hide");